Refactor analytics callbacks to use controllers
