+++ analytics/controllers/__init__.py
@@ -1,5 +1,27 @@
 """Analytics controller utilities."""
 
+from .deep_analytics_logic import (
+    dispatch_analysis,
+    get_data_sources,
+    get_initial_message_safe,
+    run_quality_analysis,
+    run_service_analysis,
+    run_suggests_analysis,
+    run_unique_patterns_analysis,
+    update_status_alert,
+)
+from .realtime_ws import RealTimeWebSocketController
 from .unified_controller import UnifiedAnalyticsController
 
-__all__ = ["UnifiedAnalyticsController"]
+__all__ = [
+    "UnifiedAnalyticsController",
+    "RealTimeWebSocketController",
+    "dispatch_analysis",
+    "run_quality_analysis",
+    "run_service_analysis",
+    "run_suggests_analysis",
+    "run_unique_patterns_analysis",
+    "update_status_alert",
+    "get_data_sources",
+    "get_initial_message_safe",
+]
